refactor: Some refactor and upgrade to yggdrasil 0.3.8

+++ app/src/main/java/io/github/chronosx88/yggdrasil/models/config/Config.kt
@@ -16,7 +16,7 @@ data class Config (
 	@SerializedName("linkLocalTCPPort") val linkLocalTCPPort : Int,
 	@SerializedName("ifName") val ifName : String,
 	@SerializedName("ifTAPMode") val ifTAPMode : Boolean,
-	@SerializedName("ifMTU") val ifMTU : Int,
+	@SerializedName("ifMTU") val ifMTU : Long,
 	@SerializedName("sessionFirewall") val sessionFirewall : SessionFirewall,
 	@SerializedName("tunnelRouting") val tunnelRouting : TunnelRouting,
 	@SerializedName("switchOptions") val switchOptions : SwitchOptions,
